lib/protoparser/datadog: follow-up after 543f218fe96574b9b2189c8350bb09afa349e3bb
* prevent /api/v1 from panic on parsing rows
* add tests for Extract function for v1 and v2 api's
* separate request types in different pools to prevent different objects mixing
* add changelog line

https://github.com/VictoriaMetrics/VictoriaMetrics/commit/543f218fe96574b9b2189c8350bb09afa349e3bb

+++ lib/protoparser/datadog/api/sketches/beta/api.go
@@ -25,7 +25,7 @@ func (r *Request) Extract(fn func(prompbmarshal.TimeSeries) error, sanitizeFn fu
 	for _, sketch := range r.SketchPayload.Sketches {
 		sketchSeries := make([]prompbmarshal.TimeSeries, 5)
 		for _, point := range sketch.Dogsketches {
-			timestamp := int64(point.Ts * 1000)
+			timestamp := point.Ts * 1000
 			updateSeries(sketchSeries, sanitizeFn(sketch.Metric), timestamp, map[string]float64{
 				"max": point.Max,
 				"min": point.Min,
@@ -35,7 +35,7 @@ func (r *Request) Extract(fn func(prompbmarshal.TimeSeries) error, sanitizeFn fu
 			})
 		}
 		for _, point := range sketch.Distributions {
-			timestamp := int64(point.Ts * 1000)
+			timestamp := point.Ts * 1000
 			updateSeries(sketchSeries, sanitizeFn(sketch.Metric), timestamp, map[string]float64{
 				"max": point.Max,
 				"min": point.Min,
